# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Staff API",
        "description": "A simple example Laravel API to manage staff records",
        "contact": {
            "name": "Warren Nelson",
            "email": "warren.nelson@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/staff": {
            "get": {
                "summary": "Lists all staff members",
                "operationId": "a9eb94a978618740b62aa58538ee1785",
                "responses": {
                    "200": {
                        "description": "A list of staff members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Staff"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Creates a new staff member",
                "operationId": "06a6bc7e0f0ffe4c19f67b8d70b58928",
                "requestBody": {
                    "description": "Staff member to add",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Staff"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Staff member created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Staff"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/staff/{id}": {
            "get": {
                "summary": "Retrieves a specific staff member by id",
                "operationId": "97c1bbd48e5834aab3b87b709f2ef517",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the staff member to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff member retrieved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Staff"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Updates a specific staff member by id",
                "operationId": "de68b1f20368fb2ad9466d2926618309",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the staff member to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data to update the staff member with",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Staff"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff member updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Staff"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deletes a specific staff member by id",
                "operationId": "3ef08e772902ed79669f197dce0f86a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the staff member to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Staff member deleted"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Staff": {
                "title": "Staff Model",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "email": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "squad": {
                        "type": "string"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "notes": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}